Remove stream framework from Dgraph.
